refactor logger out to reduce unnecessary repetition/complexity in already complex files
